make "opened" property state match show() and hide() method calls
The opened property stays true after a call to hide() and stays false after a call to show(). Adding opened = false and opened = true to the hide and show functions has the side effect of calling the _openedChanged observer which would call the hide or show one more time. Putting the logic of hiding and showing in the observer avoids this problem and the hide and show methods now leave opened in a matching state. 

The documentation says "Use opened or toggle() to show/hide the content" and doesn't mention hide or show so maybe the intention was not to call these directly. This change allows show and hide to be called as well.